Reverse clock polarity.
It's more convenient to have the registers sample the input while the
clock is high, and latch it to the output on the falling edge. That way,
a register whose clock input is kept low or Z will not change (otherwise,
we would have to keep idle registers' clock inputs high).